fix(data-warehouse): update IngestionOutputs import after module move
master relocated ingestion-outputs from src/ingestion/outputs/ to
src/common/outputs/. Point the warehouse webhook status service (and its
test) at the new ~/common/outputs/ingestion-outputs path so the nodejs
build compiles.

Generated-By: PostHog Code
Task-Id: fb7cf23b-91d0-499e-ab76-df9a61bd4ea4
