+++ nodejs/src/cdp/services/warehouse/warehouse-webhook-status.service.test.ts
@@ -1,4 +1,5 @@
-import { IngestionOutputs } from '../../../ingestion/outputs/ingestion-outputs'
+import { IngestionOutputs } from '~/common/outputs/ingestion-outputs'
+
 import { parseJSON } from '../../../utils/json-parse'
 import { WAREHOUSE_WEBHOOK_DELIVERY_STATUS_OUTPUT, WarehouseWebhookDeliveryStatusOutput } from '../../outputs/outputs'
 import { CyclotronJobInvocationResult } from '../../types'
